refactor: migrate seller dashboard API, contracts, and query hooks from auctions to seller feature module
