+++ src/mazadzone-client/src/features/auctions/api/auction.api.ts
@@ -11,10 +11,6 @@ import type {
   GetAuctionsQueryParams,
   PagedListOfAuctionsListDto,
   CategoryDto,
-  SellerAuctionsResponse,
-  SellerOrdersResponse,
-  SellerFinancialsResponse,
-  SellerDashboardQueryParams,
 } from "./auction.contracts";
 
 /**
@@ -95,41 +91,5 @@ export async function getRootCategories(): Promise<CategoryDto[]> {
  */
 export async function getCategoryTree(): Promise<CategoryDto[]> {
   const response = await api.get<CategoryDto[]>("/categories/tree");
-  return response.data;
-}
-
-/**
- * Retrieve seller dashboard auctions.
- */
-export async function getSellerDashboardAuctions(
-  params?: SellerDashboardQueryParams,
-): Promise<SellerAuctionsResponse> {
-  const response = await api.get<SellerAuctionsResponse>("/seller-dashboard/auctions", {
-    params,
-  });
-  return response.data;
-}
-
-/**
- * Retrieve seller dashboard orders.
- */
-export async function getSellerDashboardOrders(
-  params?: SellerDashboardQueryParams,
-): Promise<SellerOrdersResponse> {
-  const response = await api.get<SellerOrdersResponse>("/seller-dashboard/orders", {
-    params,
-  });
-  return response.data;
-}
-
-/**
- * Retrieve seller dashboard financials.
- */
-export async function getSellerDashboardFinancials(
-  params?: SellerDashboardQueryParams,
-): Promise<SellerFinancialsResponse> {
-  const response = await api.get<SellerFinancialsResponse>("/seller-dashboard/financials", {
-    params,
-  });
   return response.data;
 }
