+++ src/mazadzone-client/src/features/seller/api/index.ts
@@ -2,3 +2,6 @@ export * from "./seller.api";
 export * from "./seller.queries";
 export * from "./seller.mutations";
 export * from "./seller.contracts";
+export * from "./seller-dashboard.contracts";
+export * from "./seller-dashboard.api";
+export * from "./seller-dashboard.queries";
