+++ src/mazadzone-client/src/features/seller/components/seller-dashboard/orders/SellerOrdersTableRow.tsx
@@ -4,7 +4,7 @@ import { format } from "date-fns";
 import { Badge } from "@/components/ui/badge";
 import { formatCurrency } from "@/utils/currency.utils";
 import { cn } from "@/lib/utils";
-import type { SellerOrderSummaryDto } from "@/features/auctions";
+import type { SellerOrderSummaryDto } from "@/features/seller";
 
 interface SellerOrdersTableRowProps {
   order: SellerOrderSummaryDto;
